# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Informing Science API",
    "description": "API documentation for Informing Science.",
    "version": "1.0.0"
  },
  "host": "localhost:4401",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/user/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "personalName": {
                  "example": "any"
                },
                "middleInitial": {
                  "example": "any"
                },
                "familyName": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "country": {
                  "example": "any"
                },
                "affiliationUniversity": {
                  "example": "any"
                },
                "department": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/user/update": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "personalTitle": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                },
                "personalName": {
                  "example": "any"
                },
                "middleInitial": {
                  "example": "any"
                },
                "familyName": {
                  "example": "any"
                },
                "gender": {
                  "example": "any"
                },
                "profilePic": {
                  "example": "any"
                },
                "primaryTelephone": {
                  "example": "any"
                },
                "secondaryTelephone": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "stateProvince": {
                  "example": "any"
                },
                "postalCode": {
                  "example": "any"
                },
                "country": {
                  "example": "any"
                },
                "countryCode": {
                  "example": "any"
                },
                "affiliationUniversity": {
                  "example": "any"
                },
                "department": {
                  "example": "any"
                },
                "positionTitle": {
                  "example": "any"
                },
                "orcid": {
                  "example": "any"
                },
                "resume": {
                  "example": "any"
                },
                "bio": {
                  "example": "any"
                },
                "note": {
                  "example": "any"
                },
                "websiteUrl": {
                  "example": "any"
                },
                "allowProfile": {
                  "example": "any"
                },
                "unsubscribe": {
                  "example": "any"
                },
                "receivePrimaryEmail": {
                  "example": "any"
                },
                "receiveReminderEmail": {
                  "example": "any"
                },
                "isPendingAuthor": {
                  "example": "any"
                },
                "isDuplicate": {
                  "example": "any"
                },
                "socialMedia": {
                  "example": "any"
                },
                "receiveSecondaryEmail": {
                  "example": "any"
                },
                "isiPositions": {
                  "example": "any"
                },
                "isReviewerEditorOfMonth": {
                  "example": "any"
                },
                "ofTheMonth": {
                  "example": "any"
                },
                "testimonial": {
                  "example": "any"
                },
                "memberUntil": {
                  "example": "any"
                },
                "membershipTypes": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/user/profile": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/user/list": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isiPosition",
            "in": "query",
            "type": "string"
          },
          {
            "name": "ofTheMonthType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "text",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          },
          {
            "name": "role",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isDuplicate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/user/forgot/password": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "newPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/user/change/password": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "oldPassword": {
                  "example": "any"
                },
                "newPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/user/logout": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/user/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}